+++ NEWS.md
@@ -1,5 +1,7 @@
 # teal.widgets 0.1.0.9002
 
+### New features
+* Added a new widget `draggable_buckets`.
 ### Enhancements
 
 * The `option` `teal.plot_dpi` controls the `dpi` for rendering and downloading plots in `plot_with_settings`. If `option` not used then a default of `72dpi` is used